pyproject.toml
[project]
name = "my-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.85.0",
    "beautifulsoup4>=4.14.3",
    "gitpython>=3.1.46",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.4.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.14.5",
    "ruff>=0.15.12",
    "sentence-transformers>=5.3.0",
    "tavily-python>=0.7.24",
]

[tool.mypy]
[[tool.mypy.overrides]]
module = ["ragas.*"]
ignore_missing_imports = true
